Extract wizard out of auth component
