+++ src/Components/Authorization/Mobile/__stories__/RegisterForm.story.tsx
@@ -6,7 +6,7 @@ import { MobileRegisterForm } from "../RegisterForm"
 storiesOf("Components/Auth/Mobile", module).add("RegisterForm", () => (
   <MobileRegisterForm
     values={{}}
-    handleSubmit={() => console.log("called")}
-    handleChangeMode={() => mode => console.log("ok")}
+    handleSubmit={() => null}
+    handleChangeMode={() => mode => null}
   />
 ))
